Storybook 7 beta (#1285)
* test storybook 7 beta

* update apikey in test.yml

* update test.yml

* update versions test

* set JS key when running tests

* rever test.yml and main.js

* revert back to storybook 6

* add support for sb 7 in test versions test

* spawn storybook process using sb dev and render stories usign the new api

* small fix

* small fix

* improve the logic to determine storybook version before starting storybook

* remove extra brackets

* use sb binary instead of command

* undo not needed changes

* add waitBeforeCapture to storybook versions

* remove env var from test in scripts

* copyStoriesToVersionDir, check for version 7 in a better way

* update pending changes

* update loginForm.stories

* small fix

* send showLogs to dom-snapshot

* use __STORYBOOK_ADDONS_PREVIEW for to get addons for storybook 7

* simplify storybook fixtures and fix lint

* storybook versions test, consolidate the baselines of all different storybook versions

* fix lint

* listen on different error render events and add a test

* update didTestPass to return false with any type of error

* rename var to make code more readable

* test storybook with errors on storybook v6.5 and v7

* small fix

* remove unnecessary condition

* add a test for storyThrewException event

* storybook: run all tests in test:mocha and update API key to Eyes SDK team

* stabilize concurrency test

* fix lint, support node 18

* fix lint, support node 18

---------

+++ js/packages/eyes-storybook/test/fixtures/storybook-versions/mochaHooks.js
@@ -13,7 +13,7 @@ exports.mochaHooks = {
     const versionDir = path.resolve(__dirname, `./${storybookVersion}`);
     fs.readdirSync(versionDir).map(fileName => {
       const filePath = path.join(versionDir, fileName);
-      if (fs.existsSync(filePath) && fs.lstatSync(filePath).isDirectory()) {
+      if (fs.existsSync(filePath) && fs.lstatSync(filePath).isDirectory() && !filePath.includes('.storybook')) {
         fs.rmSync(filePath, { recursive: true })
       }
     });
